Aip creation fix

+++ workers/tasks.py
@@ -532,7 +532,8 @@ class SIPRestructuring(DefaultTask):
                 tl.addinfo("Restructuring content of package: %s" % str(delivery))
 
                 # TODO: maybe remove the state.xml already during SIP packaging
-                os.remove(os.path.join(str(delivery), 'state.xml'))
+                if os.path.exists(os.path.join(str(delivery), 'state.xml')):
+                    os.remove(os.path.join(str(delivery), 'state.xml'))
 
                 fs_childs =  os.listdir(str(delivery))
                 for fs_child in fs_childs:
@@ -572,17 +573,20 @@ class SIPValidation(DefaultTask):
                 tl.adderr(("%s missing: %s" % (descr, os.path.abspath(f))))
 
         submission_path = os.path.join(task_context.path, "submission")
-        check_file("SIP METS file", os.path.join(submission_path, "METS.xml"))
+        check_file("submission METS file", os.path.join(submission_path, "METS.xml"))
         #check_file("Data directory", os.path.join(reps_path, "data"))
 
         #check_file("Documentation directory", os.path.join(path, "documentation"))
         #check_file("Metadata directory", os.path.join(path, "metadata"))
-        representations_path = os.path.join(task_context.path, "submission/representations")
-        for name in os.listdir(representations_path):
-            rep_path = os.path.join(representations_path, name)
-            if os.path.isdir(rep_path):
-                mets_validator = MetsValidation(rep_path)
-                valid = mets_validator.validate_mets(os.path.join(rep_path, 'METS.xml'))
+
+        # submission is not mandatory in an AIP:
+        if os.path.exists(os.path.join(task_context.path, "submission/representations")):
+            representations_path = os.path.join(task_context.path, "submission/representations")
+            for name in os.listdir(representations_path):
+                rep_path = os.path.join(representations_path, name)
+                if os.path.isdir(rep_path):
+                    mets_validator = MetsValidation(rep_path)
+                    valid = mets_validator.validate_mets(os.path.join(rep_path, 'METS.xml'))
 
         # currently: forced valid = True, until valid mets files are created by the SIP creator!
         # valid = True
